+++ kurtosis-module/impl/module.go
@@ -3,88 +3,21 @@ package impl
 import (
 	"encoding/json"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/forkmon"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/module_io"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/cl_client_rest_client"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/lighthouse"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/lodestar"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/nimbus"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/prysm"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/teku"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el/geth"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el/nethermind"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/log_levels"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator/genesis_consts"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/transaction_spammer"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
-	"path"
-	"text/template"
 	"time"
 )
 
 const (
-	networkId = "3151908"
-
-	// ----------------------------------- Params Constants -----------------------------------------
-	defaultWaitForFinalization                                = false
-	defaultClientLogLevel      log_levels.ParticipantLogLevel = log_levels.ParticipantLogLevel_Info
-	// --------------------------------- End Params Constants ---------------------------------------
-
-	// ----------------------------------- Prelaunch Data Constants -----------------------------------------
-	// The number of validator keys that will be preregistered inside the CL genesis file when it's created
-	numValidatorsToPreregister = 100
-
-	// Seems to be hardcoded
-	slotsPerEpoch = uint32(32)
-
-	// If we drop this, things start to behave strangely, with slots that are of variable time lengths
-	secondsPerSlot = uint32(12)
-
-	// Altair must happen before merge fork must happen before terminal_total_difficulty is hit
-	// See also: https://notes.ethereum.org/@ExXcnR0-SJGthjz1dwkA1A/H1MSKgm3F
-	altairForkEpoch = uint64(1)  // Set per Parithosh's recommendation
-	mergeForkEpoch = uint64(2)   // Set per Parithosh's recommendation
-
-	// Once the total difficulty of all mined blocks crosses this threshold, the merge will be initiated
-	// Must happen after the merge fork epoch on the Beacon chain
-	totalTerminalDifficulty  = uint64(60000000)
-
-	// This menmonic will a) be used to create keystores for all the types of validators that we have and b) be used to generate a CL genesis.ssz that has the children
-	//  validator keys already preregistered as validators
-	// See also:
-	preregisteredValidatorKeysMnemonic = "giant issue aisle success illegal bike spike question tent bar rely arctic volcano long crawl hungry vocal artwork sniff fantasy very lucky have athlete"
-
-	// TODO Clarify what units these are
-	genesisDelay = 0
-
-	depositContractAddress = "0x4242424242424242424242424242424242424242"
-	// --------------------------------- End Genesis Config Constants ----------------------------------------
-
-	// ----------------------------------- Static File Constants -----------------------------------------
-	staticFilesDirpath                    = "/static-files"
-
-	// Geth + CL genesis generation
-	genesisGenerationConfigDirpath = staticFilesDirpath + "/genesis-generation-config"
-
-	elGenesisGenerationConfigDirpath = genesisGenerationConfigDirpath + "/el"
-	gethGenesisGenerationConfigYmlTemplateFilepath = elGenesisGenerationConfigDirpath + "/geth-genesis-config.yaml.tmpl"
-	nethermindGenesisGenerationJsonTemplateFilepath = elGenesisGenerationConfigDirpath + "/nethermind-genesis.json.tmpl"
-
-	clGenesisGenerationConfigDirpath = genesisGenerationConfigDirpath + "/cl"
-	clGenesisGenerationConfigYmlTemplateFilepath = clGenesisGenerationConfigDirpath + "/config.yaml.tmpl"
-	clGenesisGenerationMnemonicsYmlTemplateFilepath = clGenesisGenerationConfigDirpath + "/mnemonics.yaml.tmpl"
-
-	// Prysm
-	prysmPasswordTxtTemplateFilepath = staticFilesDirpath + "/prysm-password.txt.tmpl"
-
-	// Forkmon config
-	forkmonConfigTemplateFilepath = staticFilesDirpath + "/forkmon-config/config.toml.tmpl"
-	// --------------------------------- End Static File Constants ----------------------------------------
-
 	responseJsonLinePrefixStr = ""
 	responseJsonLineIndentStr = "  "
 
@@ -100,30 +33,6 @@ const (
 	extraDelayBeforeSlotCountStartsIncreasing = 4 * time.Minute
 )
 
-var defaultParticipants = []*ParticipantParams{
-	{
-		ELClientType: participant_network.ParticipantELClientType_Geth,
-		CLClientType: participant_network.ParticipantCLClientType_Nimbus,
-	},
-}
-
-type ParticipantParams struct {
-	ELClientType participant_network.ParticipantELClientType `json:"el"`
-	CLClientType participant_network.ParticipantCLClientType `json:"cl"`
-}
-type ExecuteParams struct {
-	// Participants
-	Participants []*ParticipantParams	`json:"participants"`
-
-	WaitForFinalization bool	`json:"waitForFinalization"`
-
-	// The log level that the started clients should log at
-	ClientLogLevel log_levels.ParticipantLogLevel `json:"logLevel"`
-}
-
-type ExecuteResponse struct {
-	ForkmonPublicURL string	`json:"forkmonUrl"`
-}
 
 type Eth2KurtosisModule struct {
 }
@@ -134,38 +43,30 @@ func NewEth2KurtosisModule() *Eth2KurtosisModule {
 
 func (e Eth2KurtosisModule) Execute(enclaveCtx *enclaves.EnclaveContext, serializedParams string) (serializedResult string, resultError error) {
 	logrus.Info("Deserializing execute params...")
-	paramsObj, err := deserializeAndValidateParams(serializedParams)
+	paramsObj, err := module_io.DeserializeAndValidateParams(serializedParams)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "An error occurred deserializing & validating the params")
 	}
+	networkParams := paramsObj.Network
 	numParticipants := uint32(len(paramsObj.Participants))
 	logrus.Info("Successfully deserialized execute params")
 
-	logrus.Info("Generating prelaunch data...")
-	genesisUnixTimestamp := time.Now().Unix()
-	gethGenesisConfigTemplate, err := parseTemplate(gethGenesisGenerationConfigYmlTemplateFilepath)
+	logrus.Info("Creating prelaunch data generator...")
+	prelaunchDataGeneratorCtx, err := prelaunch_data_generator.LaunchPrelaunchDataGenerator(
+		enclaveCtx,
+		networkParams.NetworkID,
+		networkParams.DepositContractAddress,
+		networkParams.TotalTerminalDifficulty,
+	)
 	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the Geth genesis generation config YAML template")
+		return "", stacktrace.Propagate(err, "An error occurred launching the prelaunch data-generating container")
 	}
-	clGenesisConfigTemplate, err := parseTemplate(clGenesisGenerationConfigYmlTemplateFilepath)
-	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the CL genesis generation config YAML template")
-	}
-	clGenesisMnemonicsYmlTemplate, err := parseTemplate(clGenesisGenerationMnemonicsYmlTemplateFilepath)
-	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the CL mnemonics YAML template")
-	}
-	nethermindGenesisJsonTemplate, err := parseTemplate(nethermindGenesisGenerationJsonTemplateFilepath)
-	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the Nethermind genesis json template")
-	}
-	prysmPasswordTxtTemplate, err := parseTemplate(prysmPasswordTxtTemplateFilepath)
-	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the Prysm password txt template")
-	}
+	logrus.Info("Successfully created prelaunch data generator")
+
+	/*
 	prelaunchData, err := prelaunch_data_generator.GeneratePrelaunchData(
 		enclaveCtx,
-		gethGenesisConfigTemplate,
+		chainspecAndGethGenesisGenerationConfigTemplate,
 		nethermindGenesisJsonTemplate,
 		clGenesisConfigTemplate,
 		clGenesisMnemonicsYmlTemplate,
@@ -181,50 +82,11 @@ func (e Eth2KurtosisModule) Execute(enclaveCtx *enclaves.EnclaveContext, seriali
 		totalTerminalDifficulty,
 		preregisteredValidatorKeysMnemonic,
 	)
+
 	if err != nil {
 		return "", stacktrace.Propagate(err, "An error occurred launching the Ethereum genesis generator Service")
 	}
-	logrus.Info("Successfully generated prelaunch data")
-
-	logrus.Info("Creating EL & CL client launchers...")
-	elClientLaunchers := map[participant_network.ParticipantELClientType]el.ELClientLauncher{
-		participant_network.ParticipantELClientType_Geth: geth.NewGethELClientLauncher(
-			prelaunchData.GethELGenesisJsonFilepathOnModuleContainer,
-			genesis_consts.PrefundedAccounts,
-		),
-		participant_network.ParticipantELClientType_Nethermind: nethermind.NewNethermindELClientLauncher(
-			prelaunchData.NethermindGenesisJsonFilepathOnModuleContainer,
-			totalTerminalDifficulty,
-		),
-	}
-	clGenesisPaths := prelaunchData.CLGenesisPaths
-	clClientLaunchers := map[participant_network.ParticipantCLClientType]cl.CLClientLauncher{
-		participant_network.ParticipantCLClientType_Teku: teku.NewTekuCLClientLauncher(
-			clGenesisPaths.GetConfigYMLFilepath(),
-			clGenesisPaths.GetGenesisSSZFilepath(),
-			numParticipants,
-		),
-		participant_network.ParticipantCLClientType_Nimbus: nimbus.NewNimbusLauncher(
-			clGenesisPaths.GetParentDirpath(),
-		),
-		participant_network.ParticipantCLClientType_Lodestar: lodestar.NewLodestarClientLauncher(
-			clGenesisPaths.GetConfigYMLFilepath(),
-			clGenesisPaths.GetGenesisSSZFilepath(),
-			numParticipants,
-		),
-		participant_network.ParticipantCLClientType_Lighthouse: lighthouse.NewLighthouseCLClientLauncher(
-			clGenesisPaths.GetParentDirpath(),
-			numParticipants,
-		 ),
-		participant_network.ParticipantCLClientType_Prysm: prysm.NewPrysmCLCLientLauncher(
-			clGenesisPaths.GetConfigYMLFilepath(),
-			clGenesisPaths.GetGenesisSSZFilepath(),
-			prelaunchData.KeystoresGenerationResult.PrysmPassword,
-			prysmPasswordTxtTemplate,
-			numParticipants,
-		),
-	}
-	logrus.Info("Successfully created EL & CL client launchers")
+	*/
 
 	logrus.Infof("Adding %v participants logging at level '%v'...", numParticipants, paramsObj.ClientLogLevel)
 	allParticipantSpecs := []*participant_network.ParticipantSpec{}
@@ -298,7 +160,7 @@ func (e Eth2KurtosisModule) Execute(enclaveCtx *enclaves.EnclaveContext, seriali
 		logrus.Info("First finalized epoch occurred successfully")
 	}
 
-	responseObj := &ExecuteResponse{
+	responseObj := &module_io.ExecuteResponse{
 		ForkmonPublicURL: forkmonPublicUrl,
 	}
 	responseStr, err := json.MarshalIndent(responseObj, responseJsonLinePrefixStr, responseJsonLineIndentStr)
@@ -309,52 +171,6 @@ func (e Eth2KurtosisModule) Execute(enclaveCtx *enclaves.EnclaveContext, seriali
 	return string(responseStr), nil
 }
 
-func deserializeAndValidateParams(paramsStr string) (*ExecuteParams, error) {
-	paramsObj := &ExecuteParams{
-		Participants:        defaultParticipants,
-		WaitForFinalization: defaultWaitForFinalization,
-		ClientLogLevel:      defaultClientLogLevel,
-	}
-	if err := json.Unmarshal([]byte(paramsStr), paramsObj); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred deserializing the serialized params")
-	}
-	if _, found := log_levels.ValidParticipantLogLevels[paramsObj.ClientLogLevel]; !found {
-		return nil, stacktrace.NewError("Unrecognized client log level '%v'", paramsObj.ClientLogLevel)
-	}
-	if len(paramsObj.Participants) == 0 {
-		return nil, stacktrace.NewError("At least one participant is required")
-	}
-	for idx, participant := range paramsObj.Participants {
-		if idx == 0 && participant.ELClientType == participant_network.ParticipantELClientType_Nethermind {
-			return nil, stacktrace.NewError("Cannot use a Nethermind client for the first participant because Nethermind clients don't mine on Eth1")
-		}
-
-		elClientType := participant.ELClientType
-		if _, found := participant_network.ValidParticipantELClientTypes[elClientType]; !found {
-			return nil, stacktrace.NewError("Participant %v declares unrecognized EL client type '%v'", idx, elClientType)
-		}
-
-		clClientType := participant.CLClientType
-		if _, found := participant_network.ValidParticipantCLClientTypes[clClientType]; !found {
-			return nil, stacktrace.NewError("Participant %v declares unrecognized CL client type '%v'", idx, clClientType)
-		}
-	}
-	return paramsObj, nil
-}
-
-func parseTemplate(filepath string) (*template.Template, error) {
-	tmpl, err := template.New(
-		// For some reason, the template name has to match the basename of the file:
-		//  https://stackoverflow.com/questions/49043292/error-template-is-an-incomplete-or-empty-template
-		path.Base(filepath),
-	).ParseFiles(
-		filepath,
-	)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred parsing template file '%v'", filepath)
-	}
-	return tmpl, nil
-}
 
 func waitUntilFirstFinalizedEpoch(restClient *cl_client_rest_client.CLClientRESTClient) error {
 	// If we wait long enough that we might be in this epoch, we've waited too long - finality should already have happened