+++ kurtosis-module/impl/participant_network/el/el_client_launcher.go
@@ -1,7 +1,7 @@
 package el
 
 import (
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/log_levels"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/module_io"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 )
@@ -10,7 +10,7 @@ type ELClientLauncher interface {
 	Launch(
 		enclaveCtx *enclaves.EnclaveContext,
 		serviceId services.ServiceID,
-		loglevel log_levels.ParticipantLogLevel,
+		loglevel module_io.ParticipantLogLevel,
 		networkId string,
 		// If nil, then the node will be launched as a bootnode
 		bootnodeContext *ELClientContext,