+++ docs/changelog.md
@@ -2,6 +2,7 @@
 ### Features
 * Added Prysm CL (beacon, validator) Launcher
 * Made EL & CL client log levels configurable as a module param, `logLevel`
+* Added self-documenting code for module params
 
 ### Changes
 * The `WaitForBeaconClientAvailability` method also checks if the returned status is READY, which means the node is synced